requirements.txt
aiohttp==3.8.5
aiosignal==1.3.1
argcomplete==1.10.3
async-timeout==4.0.2
attrs==23.1.0
beautifulsoup4==4.8.2
blinker==1.6.2
certifi==2023.7.22
cffi==1.15.1
chardet==3.0.4
charset-normalizer==3.2.0
click==8.1.6
compressed-rtf==1.0.6
contourpy==1.1.0
cryptography==41.0.2
cycler==0.11.0
dataclasses-json==0.5.14
Deprecated==1.2.14
docx2txt==0.8
ebcdic==1.1.1
extract-msg==0.28.7
faiss-cpu==1.7.4
filelock==3.12.2
Flask==2.3.2
Flask-DotEnv==0.1.2
Flask-WTF==1.1.1
fonttools==4.42.0
frozenlist==1.4.0
fsspec==2023.6.0
gcloud==0.18.3
googleapis-common-protos==1.59.1
greenlet==2.0.2
httplib2==0.22.0
huggingface-hub==0.16.4
idna==3.4
IMAPClient==2.1.0
itsdangerous==2.1.2
Jinja2==3.1.2
jwcrypto==1.5.0
kiwisolver==1.4.4
langchain==0.0.190
lxml==4.9.3
MarkupSafe==2.1.3
marshmallow==3.20.1
matplotlib==3.7.2
multidict==6.0.4
mypy-extensions==1.0.0
numexpr==2.8.5
numpy==1.25.2
oauth2client==4.1.3
olefile==0.46
openapi-schema-pydantic==1.2.4
packaging==23.1
pandas==2.0.3
pdfminer.six==20191110
Pillow==10.0.0
pip==23.1.2
protobuf==4.24.0rc2
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
pycryptodome==3.18.0
pydantic==1.10.12
pyparsing==3.0.9
pypdf==3.15.1
Pyrebase4==4.7.1
python-dateutil==2.8.2
python-jwt==4.0.0
python-pptx==0.6.21
pytz==2023.3
PyYAML==6.0.1
regex==2023.8.8
requests==2.29.0
requests-toolbelt==0.10.1
rsa==4.9
safetensors==0.3.2
setuptools==65.5.0
six==1.12.0
sortedcontainers==2.4.0
soupsieve==2.4.1
SQLAlchemy==2.0.19
tenacity==8.2.2
textract==1.6.5
tiktoken==0.4.0
tokenizers==0.13.3
tqdm==4.65.0
transformers==4.31.0
typing_extensions==4.7.1
typing-inspect==0.9.0
tzdata==2023.3
tzlocal==5.0.1
urllib3==1.26.16
Werkzeug==2.3.6
wrapt==1.15.0
WTForms==3.0.1
xlrd==1.2.0
XlsxWriter==3.1.2
yarl==1.9.2
